add ros2 pkg executables
