+++ ros2pkg/setup.py
@@ -32,6 +32,7 @@ The package provides the pkg command for the ROS 2 command line tools.""",
             'ros2pkg.verb = ros2pkg.verb:VerbExtension',
         ],
         'ros2pkg.verb': [
+            'executables = ros2pkg.verb.executables:ExecutablesVerb',
             'list = ros2pkg.verb.list:ListVerb',
             'prefix = ros2pkg.verb.prefix:PrefixVerb',
         ],
